requirements.txt
numpy
feedparser
python-dotenv
supabase
scikit-learn
google-genai
requests
